Split Ns Perfo SOL005.
