+++ MANO-API/src/main/java/com/ubiqube/etsi/mano/controller/nsperfo/sol005/NsPerfoSol005Api.java
@@ -1,22 +1,10 @@
 package com.ubiqube.etsi.mano.controller.nsperfo.sol005;
 
-import java.util.ArrayList;
-import java.util.List;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.ubiqube.etsi.mano.model.nsperfo.sol005.CreatePmJobRequest;
-import com.ubiqube.etsi.mano.model.nsperfo.sol005.CreateThresholdRequest;
 import com.ubiqube.etsi.mano.model.nsperfo.sol005.PerformanceInformationAvailableNotification;
-import com.ubiqube.etsi.mano.model.nsperfo.sol005.PmJobsPmJobIdReportsReportIdGetResponse;
-import com.ubiqube.etsi.mano.model.nsperfo.sol005.PmJobsPostResponse;
-import com.ubiqube.etsi.mano.model.nsperfo.sol005.SubscriptionsPostQuery;
-import com.ubiqube.etsi.mano.model.nsperfo.sol005.SubscriptionsPostResponse;
 import com.ubiqube.etsi.mano.model.nsperfo.sol005.ThresholdCrossedNotification;
-import com.ubiqube.etsi.mano.model.nsperfo.sol005.ThresholdsPostResponse;
 
 import io.swagger.annotations.Api;
 
@@ -36,195 +24,6 @@ import io.swagger.annotations.Api;
 @RequestMapping("/sol005/nspm/v1")
 @Api(value = "/")
 public class NsPerfoSol005Api implements NsPerfoSol005 {
-	/**
-	 * Query PM jobs.
-	 *
-	 * \&quot;The client can use this method to retrieve information about PM
-	 * jobs\&quot;
-	 *
-	 */
-	@Override
-	public ResponseEntity<List<Object>> pmJobsGet(String accept, String contentType, String filter, String allFields, String include, String exclude, String excludeDefault) {
-		return new ResponseEntity<>(new ArrayList<>(), HttpStatus.OK);
-	}
-
-	/**
-	 * Delete a PM job.
-	 *
-	 * This method terminates an individual PM job.
-	 *
-	 */
-	@Override
-	public void pmJobsPmJobIdDelete(String pmJobId) {
-		// TODO: Implement...
-
-	}
-
-	/**
-	 * Read a single PM job.
-	 *
-	 * The client can use this method for reading an individual PM job.
-	 *
-	 */
-	@Override
-	public ResponseEntity<PmJobsPostResponse> pmJobsPmJobIdGet(String pmJobId, String accept) {
-		// TODO: Implement...
-
-		return null;
-	}
-
-	/**
-	 * Read an individual performance report.
-	 *
-	 * The client can use this method for reading an individual performance report.
-	 *
-	 */
-	@Override
-	public ResponseEntity<PmJobsPmJobIdReportsReportIdGetResponse> pmJobsPmJobIdReportsReportIdGet(String pmJobId, String reportId, String accept) {
-		// TODO: Implement...
-
-		return null;
-	}
-
-	/**
-	 * Create a PM job.
-	 *
-	 * The POST method creates a PM job. This method shall follow the provisions
-	 * specified in the Tables 7.4.2.3.1-1 and 7.4.2.3.1-2 for URI query parameters,
-	 * request and response data structures, and response codes.
-	 *
-	 */
-	@Override
-	public ResponseEntity<PmJobsPostResponse> pmJobsPost(CreatePmJobRequest createPmJobRequest, String accept, String contentType) {
-		// TODO: Implement...
-
-		return null;
-	}
-
-	/**
-	 * Query PM related subscriptions.
-	 *
-	 * The client can use this method to query the list of active subscriptions to
-	 * Performance management notifications subscribed by the client. This method
-	 * shall follow the provisions specified in the Tables 7.4.7.3.2-1 and
-	 * 7.4.7.3.2-2 for URI query parameters, request and response data structures,
-	 * and response codes.
-	 *
-	 */
-	@Override
-	public ResponseEntity<List<Object>> subscriptionsGet(String accept, String filter) {
-		return new ResponseEntity<>(new ArrayList<>(), HttpStatus.OK);
-	}
-
-	/**
-	 * Subscribe to PM notifications.
-	 *
-	 * The POST method creates a new subscription. This method shall follow the
-	 * provisions specified in the Tables 7.4.7.3.1-1 and 7.4.7.3.1-2 for URI query
-	 * parameters, request and response data structures, and response codes.
-	 * Creation of two subscription resources with the same callbackURI and the same
-	 * filter can result in performance degradation and will provide duplicates of
-	 * notifications to the OSS, and might make sense only in very rare use cases.
-	 * Consequently, the NFVO may either allow creating a subscription resource if
-	 * another subscription resource with the same filter and callbackUri already
-	 * exists (in which case it shall return the \&quot;201 Created\&quot; response
-	 * code), or may decide to not create a duplicate subscription resource (in
-	 * which case it shall return a \&quot;303 See Other\&quot; response code
-	 * referencing the existing subscription resource with the same filter and
-	 * callbackUri)
-	 *
-	 */
-	@Override
-	public ResponseEntity<SubscriptionsPostResponse> subscriptionsPost(String accept, String contentType, SubscriptionsPostQuery body) {
-		// TODO: Implement...
-
-		return null;
-	}
-
-	/**
-	 * Terminate a subscription.
-	 *
-	 * This method terminates an individual subscription. This method shall follow
-	 * the provisions specified in the Tables 7.4.8.3.5-1 and 7.4.8.3.5-2 for URI
-	 * query parameters, request and response data structures, and response codes
-	 *
-	 */
-	@Override
-	public void subscriptionsSubscriptionIdDelete(String subscriptionId, String accept) {
-		// TODO: Implement...
-
-	}
-
-	/**
-	 * Query a single PM related subscription.
-	 *
-	 * The client can use this method for reading an individual subscription about
-	 * Performance management notifications subscribed by the client. This method
-	 * shall follow the provisions specified in the Tables 7.4.8.3.2-1 and
-	 * 7.4.8.3.2-2 for URI query parameters, request and response data structures,
-	 * and response codes.
-	 *
-	 */
-	@Override
-	public ResponseEntity<SubscriptionsPostResponse> subscriptionsSubscriptionIdGet(String subscriptionId, String accept) {
-		// TODO: Implement...
-
-		return null;
-	}
-
-	/**
-	 * Query thresholds.
-	 *
-	 * The client can use this method to query information about thresholds.
-	 *
-	 */
-	@Override
-	public ResponseEntity<List<Object>> thresholdsGet(String accept, String filter) {
-		return new ResponseEntity<>(new ArrayList<>(), HttpStatus.OK);
-	}
-
-	/**
-	 * Create a threshold.
-	 *
-	 * The POST method can be used by the client to create a threshold. This method
-	 * shall follow the provisions specified in the table 7.4.5.3.1-2 for URI query
-	 * parameters, request and response data structures, and response codes.
-	 *
-	 */
-	@Override
-	public ResponseEntity<ThresholdsPostResponse> thresholdsPost(CreateThresholdRequest createThresholdRequest, String accept, String contentType) {
-		// TODO: Implement...
-
-		return null;
-	}
-
-	/**
-	 * Delete a threshold.
-	 *
-	 * This method allows to delete a threshold.
-	 *
-	 */
-	@Override
-	public void thresholdsThresholdIdDelete(String thresholdId, String accept) {
-		// TODO: Implement...
-
-	}
-
-	/**
-	 * Query a single threshold.
-	 *
-	 * The client can use this method for reading an individual threshold. This
-	 * method shall follow the provisions specified in the Tables 7.4.6.3.2-1 and
-	 * 7.4.6.3.2-2 for URI query parameters, request and response data structures,
-	 * and response codes.
-	 *
-	 */
-	@Override
-	public ResponseEntity<ThresholdsPostResponse> thresholdsThresholdIdGet(String thresholdId, String accept) {
-		// TODO: Implement...
-
-		return null;
-	}
 
 	/**
 	 * Notify about PM related events
