+++ MANO-API/src/main/java/com/ubiqube/etsi/mano/controller/nsperfo/sol005/NsPerfoSubscriptionSol005Api.java
@@ -1,0 +1,81 @@
+package com.ubiqube.etsi.mano.controller.nsperfo.sol005;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+
+import com.ubiqube.etsi.mano.model.nsperfo.sol005.SubscriptionsPostQuery;
+import com.ubiqube.etsi.mano.model.nsperfo.sol005.SubscriptionsPostResponse;
+
+public class NsPerfoSubscriptionSol005Api implements NsPerfoSubscriptionSol005 {
+	/**
+	 * Query PM related subscriptions.
+	 *
+	 * The client can use this method to query the list of active subscriptions to
+	 * Performance management notifications subscribed by the client. This method
+	 * shall follow the provisions specified in the Tables 7.4.7.3.2-1 and
+	 * 7.4.7.3.2-2 for URI query parameters, request and response data structures,
+	 * and response codes.
+	 *
+	 */
+	@Override
+	public ResponseEntity<List<Object>> subscriptionsGet(String accept, String filter) {
+		return new ResponseEntity<>(new ArrayList<>(), HttpStatus.OK);
+	}
+
+	/**
+	 * Subscribe to PM notifications.
+	 *
+	 * The POST method creates a new subscription. This method shall follow the
+	 * provisions specified in the Tables 7.4.7.3.1-1 and 7.4.7.3.1-2 for URI query
+	 * parameters, request and response data structures, and response codes.
+	 * Creation of two subscription resources with the same callbackURI and the same
+	 * filter can result in performance degradation and will provide duplicates of
+	 * notifications to the OSS, and might make sense only in very rare use cases.
+	 * Consequently, the NFVO may either allow creating a subscription resource if
+	 * another subscription resource with the same filter and callbackUri already
+	 * exists (in which case it shall return the \&quot;201 Created\&quot; response
+	 * code), or may decide to not create a duplicate subscription resource (in
+	 * which case it shall return a \&quot;303 See Other\&quot; response code
+	 * referencing the existing subscription resource with the same filter and
+	 * callbackUri)
+	 *
+	 */
+	@Override
+	public ResponseEntity<SubscriptionsPostResponse> subscriptionsPost(String accept, String contentType, SubscriptionsPostQuery body) {
+		// : Implement...
+		return null;
+	}
+
+	/**
+	 * Terminate a subscription.
+	 *
+	 * This method terminates an individual subscription. This method shall follow
+	 * the provisions specified in the Tables 7.4.8.3.5-1 and 7.4.8.3.5-2 for URI
+	 * query parameters, request and response data structures, and response codes
+	 *
+	 */
+	@Override
+	public void subscriptionsSubscriptionIdDelete(String subscriptionId, String accept) {
+		// : Implement...
+	}
+
+	/**
+	 * Query a single PM related subscription.
+	 *
+	 * The client can use this method for reading an individual subscription about
+	 * Performance management notifications subscribed by the client. This method
+	 * shall follow the provisions specified in the Tables 7.4.8.3.2-1 and
+	 * 7.4.8.3.2-2 for URI query parameters, request and response data structures,
+	 * and response codes.
+	 *
+	 */
+	@Override
+	public ResponseEntity<SubscriptionsPostResponse> subscriptionsSubscriptionIdGet(String subscriptionId, String accept) {
+		// : Implement...
+		return null;
+	}
+
+}
