+++ MANO-API/src/main/java/com/ubiqube/etsi/mano/controller/nsperfo/sol005/ThresholdsSol005Api.java
@@ -1,0 +1,68 @@
+package com.ubiqube.etsi.mano.controller.nsperfo.sol005;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+
+import com.ubiqube.etsi.mano.model.nsperfo.sol005.CreateThresholdRequest;
+import com.ubiqube.etsi.mano.model.nsperfo.sol005.ThresholdsPostResponse;
+
+public class ThresholdsSol005Api implements ThresholdsSol005 {
+
+	/**
+	 * Query thresholds.
+	 *
+	 * The client can use this method to query information about thresholds.
+	 *
+	 */
+	@Override
+	public ResponseEntity<List<Object>> thresholdsGet(String accept, String filter) {
+		return new ResponseEntity<>(new ArrayList<>(), HttpStatus.OK);
+	}
+
+	/**
+	 * Create a threshold.
+	 *
+	 * The POST method can be used by the client to create a threshold. This method
+	 * shall follow the provisions specified in the table 7.4.5.3.1-2 for URI query
+	 * parameters, request and response data structures, and response codes.
+	 *
+	 */
+	@Override
+	public ResponseEntity<ThresholdsPostResponse> thresholdsPost(CreateThresholdRequest createThresholdRequest, String accept, String contentType) {
+		// : Implement...
+
+		return null;
+	}
+
+	/**
+	 * Delete a threshold.
+	 *
+	 * This method allows to delete a threshold.
+	 *
+	 */
+	@Override
+	public void thresholdsThresholdIdDelete(String thresholdId, String accept) {
+		// : Implement...
+
+	}
+
+	/**
+	 * Query a single threshold.
+	 *
+	 * The client can use this method for reading an individual threshold. This
+	 * method shall follow the provisions specified in the Tables 7.4.6.3.2-1 and
+	 * 7.4.6.3.2-2 for URI query parameters, request and response data structures,
+	 * and response codes.
+	 *
+	 */
+	@Override
+	public ResponseEntity<ThresholdsPostResponse> thresholdsThresholdIdGet(String thresholdId, String accept) {
+		// : Implement...
+
+		return null;
+	}
+
+}
